primary changes

+++ src/ITCHParser.hpp
@@ -2,9 +2,6 @@
 #include <cstdint>
 #include <string>
 
-// ── ITCH Message Types ──────────────────────────────────
-// These are the only message types we care about
-
 enum class ITCHMessageType : char {
     ADD_ORDER           = 'A',
     ADD_ORDER_MPID      = 'F',
@@ -15,14 +12,12 @@ enum class ITCHMessageType : char {
     UNKNOWN             = '?'
 };
 
-// ── Parsed message structs ──────────────────────────────
-
 struct AddOrderMsg {
-    uint64_t order_ref;      // unique order ID
-    char     side;           // 'B' = buy, 'S' = sell
-    uint32_t shares;         // quantity
-    char     stock[9];       // ticker symbol (8 chars + null)
-    double   price;          // price (decoded from fixed point)
+    uint64_t order_ref;
+    char     side;
+    uint32_t shares;
+    char     stock[9];
+    int64_t  price;  
 };
 
 struct CancelOrderMsg {
@@ -38,15 +33,11 @@ struct ReplaceOrderMsg {
     uint64_t old_order_ref;
     uint64_t new_order_ref;
     uint32_t shares;
-    double   price;
+    int64_t  price;   
 };
-
-// ── The Parser Class ────────────────────────────────────
 
 class ITCHParser {
 public:
-    // Decode one raw ITCH message buffer
-    // Returns the message type so caller knows which struct to use
     static ITCHMessageType getType(const uint8_t* buf);
 
     static AddOrderMsg     parseAddOrder    (const uint8_t* buf);
@@ -56,13 +47,12 @@ public:
     static ReplaceOrderMsg parseReplaceOrder(const uint8_t* buf);
 
 private:
-    // ITCH is big-endian — these helpers swap bytes
     static uint16_t read16(const uint8_t* buf, int offset);
     static uint32_t read32(const uint8_t* buf, int offset);
-    static uint64_t read48(const uint8_t* buf, int offset); // 6-byte int
+    static uint64_t read48(const uint8_t* buf, int offset);
     static uint64_t read64(const uint8_t* buf, int offset);
 
-    // ITCH prices are fixed-point: divide by 10000
-    static double readPrice(const uint8_t* buf, int offset);
+    static int64_t readPrice(const uint8_t* buf, int offset);
 };
+
